Added ability to define users/groups to get `SELECT` rights and additional page properties to check for.
Now you can use `config.json` to dictate for each schema which database
users should have select rights for the tables created in the that
schema, as well as provide a list of any additional page properties you
created via Segment that should be loaded into the tables. This
required altering the functions to take additional arguments.